CI-29: changes after review
